+++ source/helpers/getPhoneInputWithCountryStateUpdateFromNewProps.js
@@ -5,6 +5,8 @@ import {
 	couldNumberBelongToCountry
 } from './phoneInputHelpers.js'
 
+import { validateE164Number } from './isE164Number.js'
+
 import getInternationalPhoneNumberPrefix from './getInternationalPhoneNumberPrefix.js'
 
 import {
@@ -136,6 +138,12 @@ export default function getPhoneInputWithCountryStateUpdateFromNewProps(props, p
 		let phoneNumber
 		let parsedCountry
 		if (newValue) {
+			// Validate that the newly-supplied `value` is in `E.164` format.
+			// Because sometimes people attempt to supply a `value` like "+1 (879) 490-8676".
+			// https://gitlab.com/catamphetamine/react-phone-number-input/-/issues/231#note_2016334796
+			if (newValue) {
+				validateE164Number(newValue)
+			}
 			phoneNumber = parsePhoneNumber(newValue, metadata)
 			const supportedCountries = getSupportedCountries(countries, metadata)
 			if (phoneNumber && phoneNumber.country) {