+++ source/usePhoneDigits.js
@@ -2,6 +2,7 @@ import { useRef, useState, useCallback, useEffect } from 'react'
 import { AsYouType, getCountryCallingCode, parseDigits } from 'libphonenumber-js/core'
 
 import getInternationalPhoneNumberPrefix from './helpers/getInternationalPhoneNumberPrefix.js'
+import { validateE164Number } from './helpers/isE164Number.js'
 
 /**
  * Returns `[phoneDigits, setPhoneDigits]`.
@@ -24,6 +25,13 @@ export default function usePhoneDigits({
 	}
 
 	const getInitialPhoneDigits = (options) => {
+		// Validate that the initially-supplied `value` is in `E.164` format.
+		// Because sometimes people attempt to supply a `value` like "+1 (879) 490-8676".
+		// https://gitlab.com/catamphetamine/react-phone-number-input/-/issues/231#note_2016334796
+		if (value) {
+			validateE164Number(value)
+		}
+
 		return getPhoneDigitsForValue(
 			value,
 			country,